Develop

+++ src/ft_nbrlen.c
@@ -6,13 +6,13 @@
 /*   By: jlagneau <jlagneau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/11 16:34:03 by jlagneau          #+#    #+#             */
-/*   Updated: 2016/09/11 16:42:50 by jlagneau         ###   ########.fr       */
+/*   Updated: 2017/04/08 12:22:29 by jlagneau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <ft_def.h>
+#include <libft.h>
 
-size_t	ft_nbrlen(int n)
+size_t				ft_nbrlen(long n)
 {
 	const t_bool	is_negative = (n < 0) ? TRUE : FALSE;
 	size_t			size;
